refactor: 일부 method를 Service로 이동
`BaseballGameController`의 일부 메서드를 `BaseballGameService`로 이동함.
추후 service의 기능 별로 service를 분리할 계획이 있음.
