+++ src/main/java/baseball/Application.java
@@ -5,6 +5,6 @@ import baseball.controller.BaseballGameController;
 public class Application {
     public static void main(String[] args) {
         BaseballGameController baseballGameController = new BaseballGameController();
-        baseballGameController.startBaseballGame();
+        baseballGameController.gameStart();
     }
 }
